TD-1555 Updates web tests relating to tracker url
