+++ DigitalLearningSolutions.Web.Tests/ViewModels/LearningMenu/DiagnosticContentViewModelTests.cs
@@ -13,6 +13,7 @@
         private IConfiguration config = null!;
 
         private const string BaseUrl = "https://example.com";
+        private const string AppRootPathUrl = "https://www.dls.nhs.uk/v2";
         private const int CustomisationId = 5;
         private const int CentreId = 6;
         private const int SectionId = 7;
@@ -24,6 +25,7 @@
         {
             config = A.Fake<IConfiguration>();
             A.CallTo(() => config["CurrentSystemBaseUrl"]).Returns(BaseUrl);
+            A.CallTo(() => config["AppRootPath"]).Returns(AppRootPathUrl);
         }
 
         [Test]
@@ -160,7 +162,7 @@
             diagnosticContentViewModel.ContentSource.Should().Be(
                 "https://www.dls.nhs.uk/CMS/CMSContent/Course119/Diagnostic/07DiagnosticTesting/itspplayer.html" +
                 "?CentreID=6&CustomisationID=5&CandidateID=8&SectionID=7&Version=55&ProgressID=9" +
-                $"&type=diag&TrackURL={BaseUrl}/tracking/tracker&objlist=[1,2,3]&plathresh=77"
+                $"&type=diag&TrackURL={AppRootPathUrl}/tracking/tracker&objlist=[1,2,3]&plathresh=77"
             );
         }
 
@@ -196,7 +198,7 @@
             diagnosticContentViewModel.ContentSource.Should().Be(
                 "https://www.dls.nhs.uk/CMS/CMSContent/Course119/Diagnostic/07DiagnosticTesting/itspplayer.html" +
                 "?CentreID=6&CustomisationID=5&CandidateID=8&SectionID=7&Version=55&ProgressID=9" +
-                $"&type=diag&TrackURL={BaseUrl}/tracking/tracker&objlist=[1,2,3,4]&plathresh=77"
+                $"&type=diag&TrackURL={AppRootPathUrl}/tracking/tracker&objlist=[1,2,3,4]&plathresh=77"
             );
         }
